src/requirements.txt
responder==2.0.5
uvicorn==0.10.9
gunicorn==20.0.4
asgi-ratelimit[full]